status/requirements.txt
aio-pika==9.5.5
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiormq==6.8.1
aiosignal==1.4.0
attrs==25.3.0
dnspython==2.7.0
exceptiongroup==1.3.0
frozenlist==1.7.0
idna==3.10
motor==3.7.1
multidict==6.6.3
pamqp==3.3.0
propcache==0.3.2
pymongo==4.13.2
python-dotenv==1.1.1
yarl==1.20.1
